+++ GHEtool/Examples/effect_of_borehole_configuration.py
@@ -2,10 +2,11 @@
 This document is an example of how the borefield configuration can influence the total borehole length and hence the cost of the borefield.
 """
 
-# import all the relevant functions
-from GHEtool import GroundData, Borefield
 import numpy as np
 import pygfunction as gt
+
+# import all the relevant functions
+from GHEtool import Borefield, GroundData
 
 
 def effect_borefield_configuration():