+++ GHEtool/VariableClasses/GFunction.py
@@ -1,7 +1,8 @@
-import pygfunction as gt
-import numpy as np
-from typing import Union, Tuple, List
 import copy
+from typing import List, Tuple, Union
+
+import numpy as np
+import pygfunction as gt
 from scipy import interpolate
 
 from .CustomGFunction import _timeValues
@@ -48,7 +49,7 @@ class FIFO:
 
     def in_fifo_list(self, value: float) -> bool:
         """
-        This function checks whether or not the value is in the fifo list, but not the last element!
+        This function checks whether the value is in the fifo list, but not the last element!
 
         Parameters
         ----------