+++ GHEtool/gui/start_gui.py
@@ -1,7 +1,7 @@
 from sys import argv
 
 
-def run(path_list=None):  # pragma: no cover
+def run(path=None):  # pragma: no cover
     from pathlib import Path
     from configparser import ConfigParser
     from ctypes import windll as ctypes_windll
@@ -10,15 +10,15 @@ def run(path_list=None):  # pragma: no cover
     from PySide6.QtWidgets import QApplication as QtWidgets_QApplication
     from PySide6.QtWidgets import QMainWindow as QtWidgets_QMainWindow
 
-    from GHEtool import FOLDER
+    from GHEtool import FOLDER, ghe_logger
     from GHEtool.gui.gui_combine_window import MainWindow
+    from GHEtool.logger.ghe_logger import console_handler
 
     # init application
     app = QtWidgets_QApplication()
     # get current version
-    path = Path(FOLDER).parent
     config = ConfigParser()
-    config.read_file(open(path.joinpath('setup.cfg'), 'r'))
+    config.read_file(open(Path(FOLDER).parent.joinpath('setup.cfg'), 'r'))
     version = config.get('metadata', 'version')
     # set version and id
     myAppID = f'GHEtool v{version}'  # arbitrary string
@@ -30,8 +30,8 @@ def run(path_list=None):  # pragma: no cover
     # init gui window
     main_window = MainWindow(window, app)
     # load file if it is in path list
-    if path_list is not None:
-        main_window.filename = ([path for path in path_list if path.endswith('.GHEtool')][0], 0)
+    if path is not None:
+        main_window.filename = (path, 0)
         main_window.fun_load_known_filename()
 
     # show window
@@ -40,6 +40,8 @@ def run(path_list=None):  # pragma: no cover
         pyi_splash.close()
     except ModuleNotFoundError:
         pass
+
+    ghe_logger.info('GHEtool loaded!')
     window.showMaximized()
     # close app
     sys_exit(app.exec())
@@ -47,4 +49,4 @@ def run(path_list=None):  # pragma: no cover
 
 if __name__ == "__main__":  # pragma: no cover
     # pass system args like a file to read
-    run(argv if len(argv) > 1 else None)
+    run([path for path in argv if path.endswith('.GHEtool')][0] if len(argv) > 1 else None)