+++ GHEtool/logger/__init__.py
@@ -1,0 +1,4 @@
+"""
+logger package
+"""
+from GHEtool.logger.ghe_logger import ghe_logger